CalibreDb: Move state type
