+++ ui/src/calibredb/components/CalibreDbProvider.tsx
@@ -3,9 +3,23 @@ import { Dispatch, ReactNode, SetStateAction, createContext } from 'react';
 import { CalibreDbMetadata } from '@/contracts/site';
 import { KeyValueOfType } from '@/utils';
 
-import { CalibreDbState } from '../hooks/useCalibreDb';
 import { useCalibreDbState } from '../hooks/useCalibreDbState';
 import { SearchParamState } from '../utils/searchParamStateMapper';
+
+export type CalibreDbState = {
+  apiPath: string;
+  title: string;
+  pages: number;
+  shuffleSeed?: number;
+  mode: string;
+  titleContains?: string;
+  author?: number;
+  format?: number;
+  bookPath?: string;
+  readStatus?: boolean;
+  sortOrder: string;
+  devices?: string[];
+};
 
 export type CalibreDbProviderValue = {
   state: CalibreDbState;
